Add UBL 2.0, 2.1 and 2.2 as backports.
This is due to limitations in maven that prevents you from depending on multiple versions of the same artifact,	this project support backports of previous versions.

This enables including multiple versions of the UBL XML schemas in your project, you can use this project for earlier versions and then https://github.com/Tradeshift/tradeshift-ubl-xsd for the latest version

Do not use this project if you only need one version of the schemas.

UBL 2.2 is added as backport for completeness and because UBL 2.3 is not that far out.